refactor: update icon component and create icon wrapper component
