+++ packages/shared/components/atoms/index.js
@@ -10,7 +10,6 @@ export { default as AmountBox } from './AmountBox.svelte'
 export { default as BalanceSummaryRow } from './BalanceSummaryRow.svelte'
 export { default as BalanceSummarySection } from './BalanceSummarySection.svelte'
 export { default as ColoredCircle } from './ColoredCircle.svelte'
-export { default as Icon } from './Icon.svelte'
 export { default as LedgerAnimation } from './LedgerAnimation.svelte'
 export { default as MediaDisplay } from './MediaDisplay.svelte'
 export { default as MenuItem } from './MenuItem.svelte'
