+++ packages/shared/components/index.js
@@ -24,6 +24,8 @@ export { default as Toast } from './Toast.svelte'
 export { default as ToastContainer } from './ToastContainer.svelte'
 export { default as Tooltip } from './Tooltip.svelte'
 export { default as Transition } from './Transition.svelte'
+export { default as Icon } from './Icon.svelte'
+export { default as IconBox } from './IconBox.svelte'
 
 export * from './atoms'
 export * from './enums'
